Renamed apps/CustomApp.py to follow snake case convention
